+++ src/google_ads_mcp/__init__.py
@@ -3,5 +3,5 @@
 Created by Akela (https://github.com/akelaonline).
 """
 
-__version__ = "0.16.7"
+__version__ = "0.16.8"
 __author__ = "Akela"